refactor: add stubs for oidc4ci flows to wallet cli
