+++ component/wallet-cli/main.go
@@ -16,13 +16,16 @@ import (
 
 func main() {
 	rootCmd := &cobra.Command{
-		Use: "wallet-cli",
+		Use:   "wallet-cli",
+		Short: "Wallet CLI",
+		Long:  "Wallet CLI is a testing tool that emulates Wallet for OIDC4VC flows.",
 		Run: func(cmd *cobra.Command, args []string) {
 			cmd.HelpFunc()(cmd, args)
 		},
 	}
 
 	rootCmd.AddCommand(cmd.NewOIDC4VPCommand())
+	rootCmd.AddCommand(cmd.NewOIDC4CICommand())
 
 	if err := rootCmd.Execute(); err != nil {
 		log.Fatalf("Failed to run wallet-cli: %s", err)
