+++ component/wallet-cli/pkg/walletrunner/wallet_runner_oidc4ci.go
@@ -1,0 +1,16 @@
+/*
+Copyright Avast Software. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package walletrunner
+
+type OIDC4CIConfig struct {
+	InitiateIssuanceURL string
+	ClientID            string
+}
+
+func (s *Service) RunOIDC4CI(config *OIDC4CIConfig) error {
+	return nil
+}
