+++ component/wallet-cli/pkg/walletrunner/wallet_runner_oidc4ci_pre_auth.go
@@ -1,0 +1,11 @@
+/*
+Copyright Avast Software. All Rights Reserved.
+
+SPDX-License-Identifier: Apache-2.0
+*/
+
+package walletrunner
+
+func (s *Service) RunOIDC4CIPreAuth(config *OIDC4CIConfig) error {
+	return nil
+}
